refactor: rename 'atoms' to 'design'
