+++ src/view/design/Caption.tsx
