+++ src/view/design/StatusBar.tsx
