+++ src/view/pages/HomeScreen.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react'
 import { Platform } from 'react-native'
-import { ScreenView } from '../atoms/ScreenView'
+import { ScreenView } from '../design/ScreenView'
 import { HomeScreenHeader, HomeScreenFAB, HomeScreenMainView } from '../templates/HomeScreenView'
 import { PublicTimeline } from '../organisms/PublicTimeline'
 import { PublicTimelineBlocProvider, usePublicTimelineBloc } from '../../hooks/usePublicTimelineBloc'
