+++ src/view/templates/AuthorizeWithCodeScreenView.tsx
@@ -2,12 +2,12 @@ import React, { useState, useCallback } from 'react'
 import { StyleSheet } from 'react-native'
 import { getStatusBarHeight } from 'react-native-status-bar-height'
 import { Title, Caption, Divider, Button, Text, Snackbar } from 'react-native-paper'
-import { ScreenView } from '../atoms/ScreenView'
+import { ScreenView } from '../design/ScreenView'
 import { Observable } from 'rxjs'
 import { useObservableEffect } from '../../hooks/useObservableEffect'
 import { useSnackBar } from '../../hooks/useSnackBar'
-import { PaperTextInput } from '../atoms/TextInput'
-import { KeyboardAvoidingView } from '../atoms/KeyboardAvoidingView'
+import { PaperTextInput } from '../design/TextInput'
+import { KeyboardAvoidingView } from '../design/KeyboardAvoidingView'
 import { Header } from 'react-navigation'
 
 export const AuthorizeWithCodeScreenView: React.FC<{
