+++ src/view/templates/FileModalScreenView/index.tsx
@@ -3,7 +3,7 @@ import { View, StatusBar, ViewStyle, Dimensions } from 'react-native'
 import { getStatusBarHeight } from 'react-native-status-bar-height'
 import { Theme, withTheme, Appbar } from 'react-native-paper'
 import { File } from '../../../models'
-import { ScreenView } from '../../atoms/ScreenView'
+import { ScreenView } from '../../design/ScreenView'
 import { ImageModal } from '../../molecules/ImageModal'
 import { VideoModal } from '../../molecules/VideoModal'
 import color from 'color'
