+++ src/view/templates/PostModalScreenView/index.tsx
@@ -3,13 +3,13 @@ import { Observable, EMPTY } from 'rxjs'
 import { useObservableEffect } from '../../../hooks/useObservableEffect'
 import { ToastAndroid, ViewStyle, View, TextStyle } from 'react-native'
 import { getStatusBarHeight } from 'react-native-status-bar-height'
-import { ScreenView } from '../../atoms/ScreenView'
+import { ScreenView } from '../../design/ScreenView'
 import { Header } from 'react-navigation'
 import { Appbar } from 'react-native-paper'
 import { useTheme } from '../../../hooks/useTheme'
-import { KeyboardAvoidingView } from '../../atoms/KeyboardAvoidingView'
-import { TextInput } from '../../atoms/TextInput'
-import { AppHeader } from '../../atoms/AppHeader'
+import { KeyboardAvoidingView } from '../../design/KeyboardAvoidingView'
+import { TextInput } from '../../design/TextInput'
+import { AppHeader } from '../../design/AppHeader'
 import { headerColor } from '../../color'
 
 type Props = Partial<{
