+++ src/view/templates/SignInScreenView.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { StyleSheet, View } from 'react-native'
 import { useTheme } from '../../hooks/useTheme'
-import { ScreenView } from '../atoms/ScreenView'
+import { ScreenView } from '../design/ScreenView'
 import { Title, Caption, Text, Divider, Button } from 'react-native-paper'
 
 export const SignInScreenView: React.FC<{
